[REFACTOR] 폴더 정리
