+++ src/main/java/org/hanihome/hanihomebe/temporaryProperty/application/TemporaryPropertyService.java
@@ -9,7 +9,7 @@ import org.hanihome.hanihomebe.item.repository.OptionItemRepository;
 import org.hanihome.hanihomebe.member.domain.Member;
 import org.hanihome.hanihomebe.member.repository.MemberRepository;
 import org.hanihome.hanihomebe.property.domain.enums.PropertySuperType;
-import org.hanihome.hanihomebe.temporaryProperty.StepValidationManager;
+import org.hanihome.hanihomebe.temporaryProperty.application.validator.StepValidationManager;
 import org.hanihome.hanihomebe.temporaryProperty.domain.TemporaryProperty;
 import org.hanihome.hanihomebe.temporaryProperty.domain.TemporaryRentProperty;
 import org.hanihome.hanihomebe.temporaryProperty.domain.TemporaryShareProperty;
