+++ src/main/java/org/hanihome/hanihomebe/temporaryProperty/application/validator/impl/ConditionStepValidator.java
@@ -1,7 +1,8 @@
-package org.hanihome.hanihomebe.temporaryProperty;
+package org.hanihome.hanihomebe.temporaryProperty.application.validator.impl;
 
 import org.hanihome.hanihomebe.global.exception.CustomException;
 import org.hanihome.hanihomebe.global.response.domain.ServiceCode;
+import org.hanihome.hanihomebe.temporaryProperty.application.validator.interfac.StepValidator;
 import org.hanihome.hanihomebe.temporaryProperty.domain.enums.TemporaryPropertyStepStatus;
 import org.hanihome.hanihomebe.temporaryProperty.web.dto.ConditionDTO;
 import org.springframework.stereotype.Component;
