+++ src/main/java/org/hanihome/hanihomebe/temporaryProperty/application/validator/impl/DescriptionStepValidator.java
@@ -1,11 +1,12 @@
-package org.hanihome.hanihomebe.temporaryProperty;
+package org.hanihome.hanihomebe.temporaryProperty.application.validator.impl;
 
 import org.hanihome.hanihomebe.global.exception.CustomException;
 import org.hanihome.hanihomebe.global.response.domain.ServiceCode;
+import org.hanihome.hanihomebe.temporaryProperty.application.validator.interfac.StepValidator;
 import org.hanihome.hanihomebe.temporaryProperty.domain.enums.TemporaryPropertyStepStatus;
 import org.hanihome.hanihomebe.temporaryProperty.web.dto.DescriptionDTO;
 
-public class DescriptionStepValidator implements StepValidator<DescriptionDTO>{
+public class DescriptionStepValidator implements StepValidator<DescriptionDTO> {
 
     @Override
     public TemporaryPropertyStepStatus getStepStatus() {
