+++ src/main/java/org/hanihome/hanihomebe/temporaryProperty/application/validator/interfac/StepValidator.java
@@ -1,6 +1,5 @@
-package org.hanihome.hanihomebe.temporaryProperty;
+package org.hanihome.hanihomebe.temporaryProperty.application.validator.interfac;
 
-import org.hanihome.hanihomebe.temporaryProperty.domain.TemporaryProperty;
 import org.hanihome.hanihomebe.temporaryProperty.domain.enums.TemporaryPropertyStepStatus;
 
 public interface StepValidator<T> {
